Move types to own file
